requirements.txt
requests==2.31.0
Flask==2.3.2
twilio==8.4.0
python-dotenv==1.0.0
openai==1.13.0
